Introduce HmacTestUtils to test the native implementation without messing around with kotlin types
