+++ ios/Sample/sampleTests/sampleTests.swift
@@ -32,8 +32,13 @@ class sampleTests: XCTestCase {
     }
     
     func testHmac256Signature() {
+        let stringToSign = "AWS4-HMAC-SHA256\n" +
+            "20150830T123600Z\n" +
+            "20150830/us-east-1/iam/aws4_request\n" +
+        "f536975d06c0309214f805bb90ccff089219ecd68b2577efef23edd43b7e1a59"
+        let signingKey = "c4afb1cc5771d871763a393e44b703571b55cc28424d1a5e86da6ed3c154a4b9"
         
-        let signature = HmacTester().signature()
+        let signature = HmacTestUtils().createSignature(stringToSign: stringToSign, signingKey: signingKey)
         
         assert(signature == "fe52b221b5173b501c9863cec59554224072ca34c1c827ec5fb8a257f97637b1")
     }
